jobs/requirements.txt
geopandas
rasterio
numpy
psutil
python-dotenv
